+++ packages/frontend/src/components/UserSettings/GithubSettingsPanel/index.tsx
@@ -1,4 +1,3 @@
-import { type ApiError, type GitRepo } from '@lightdash/common';
 import {
     Alert,
     Avatar,
@@ -18,70 +17,16 @@ import {
     IconRefresh,
     IconTrash,
 } from '@tabler/icons-react';
-import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
 import { useEffect, type FC } from 'react';
-import { lightdashApi } from '../../../api';
 import useToaster from '../../../hooks/toaster/useToaster';
 import useSearchParams from '../../../hooks/useSearchParams';
 import githubIcon from '../../../svgs/github-icon.svg';
+import {
+    useDeleteGithubInstallationMutation,
+    useGitHubRepositories,
+} from '../../common/GithubIntegration/hooks/useGithubIntegration';
 import MantineIcon from '../../common/MantineIcon';
 import { SettingsGridCard } from '../../common/Settings/SettingsCard';
-
-const getGithubRepositories = async () =>
-    lightdashApi<GitRepo[]>({
-        url: `/github/repos/list`,
-        method: 'GET',
-        body: undefined,
-    });
-
-const useGitHubRepositories = () => {
-    const { showToastApiError } = useToaster();
-
-    return useQuery<GitRepo[], ApiError>({
-        queryKey: ['github_branches'],
-        queryFn: () => getGithubRepositories(),
-        retry: false,
-        onError: ({ error }) => {
-            if (error.statusCode === 404 || error.statusCode === 401) return; // Ignore missing installation errors or unauthorized in demo
-
-            showToastApiError({
-                title: 'Failed to get GitHub integration',
-                apiError: error,
-            });
-        },
-    });
-};
-const deleteGithubInstallation = async () =>
-    lightdashApi<null>({
-        url: `/github/uninstall`,
-        method: 'DELETE',
-        body: undefined,
-    });
-
-const useDeleteGithubInstallationMutation = () => {
-    const { showToastSuccess, showToastApiError } = useToaster();
-    const queryClient = useQueryClient();
-    return useMutation<null, ApiError>(
-        ['delete_github_installation'],
-        () => deleteGithubInstallation(),
-        {
-            onSuccess: async () => {
-                await queryClient.invalidateQueries(['github_branches']);
-                showToastSuccess({
-                    title: 'GitHub integration deleted',
-                    subtitle:
-                        'You have successfully deleted your GitHub integration.',
-                });
-            },
-            onError: ({ error }) => {
-                showToastApiError({
-                    title: 'Failed to delete GitHub integration',
-                    apiError: error,
-                });
-            },
-        },
-    );
-};
 
 const GITHUB_INSTALL_URL = `/api/v1/github/install`;