+++ packages/frontend/src/components/common/PivotTable/index.tsx
@@ -33,6 +33,7 @@ import last from 'lodash/last';
 import { readableColor } from 'polished';
 import React, { useCallback, useEffect, useMemo, useRef, type FC } from 'react';
 import { getDecimalPrecision } from '../../../hooks/tableVisualization/getDataAndColumns';
+import { formatCellContent } from '../../../hooks/useColumns';
 import { getColorFromRange, isHexCodeColor } from '../../../utils/colorUtils';
 import { getConditionalRuleLabel } from '../Filters/FilterInputs/utils';
 import Table from '../LightTable';
@@ -202,9 +203,7 @@ const PivotTable: FC<PivotTableProps> = ({
                     },
                     {
                         id: col.fieldId,
-                        cell: (info) => {
-                            return info.getValue()?.value?.formatted || '-';
-                        },
+                        cell: (info) => formatCellContent(info.getValue()),
                         meta: {
                             item: item,
                             type: col.columnType,
@@ -404,7 +403,6 @@ const PivotTable: FC<PivotTableProps> = ({
                                 #
                             </Table.CellHead>
                         )}
-
                         {/* renders the title labels */}
                         {data.titleFields[headerRowIndex].map(
                             (titleField, titleFieldIndex) => {
@@ -441,7 +439,6 @@ const PivotTable: FC<PivotTableProps> = ({
                                 );
                             },
                         )}
-
                         {/* renders the header values or labels */}
                         {headerValues.map((headerValue, headerColIndex) => {
                             const isLabel = headerValue.type === 'label';
@@ -465,11 +462,10 @@ const PivotTable: FC<PivotTableProps> = ({
                                 >
                                     {isLabel
                                         ? getFieldLabel(headerValue.fieldId)
-                                        : headerValue.value.formatted}
+                                        : formatCellContent(headerValue)}
                                 </Table.CellHead>
                             ) : null;
                         })}
-
                         {/* render the total label */}
                         {hasRowTotals
                             ? data.rowTotalFields?.[headerRowIndex].map(